move module bundling to `postinstall` script
