+++ .gitignore
@@ -103,3 +103,4 @@ dist
 # TernJS port file
 .tern-port
 .DS_Store
+modules
